third_party/blink/web_tests/external/wpt/editing/plaintext-only/insertText_white-space=normal-expected.txt

This is a testharness.js-based test.
[FAIL] execCommand("insertText", false, "  ") when <b>A[]B</b>
  assert_in_array: value "<b>A  B</b>" not in array ["<b>A &nbsp;B</b>", "<b>A&nbsp; B</b>"]
[FAIL] Typing "  " when <b>A[]B</b>
  assert_in_array: value "<b>A  B</b>" not in array ["<b>A &nbsp;B</b>", "<b>A&nbsp; B</b>"]
[FAIL] execCommand("insertText", false, "  ") when <p style="white-space:normal">A[]B</p>
  assert_in_array: value "<p style=\\"white-space:normal\\">A  B</p>" not in array ["<p style=\\"white-space:normal\\">A &nbsp;B</p>", "<p style=\\"white-space:normal\\">A&nbsp; B</p>"]
[FAIL] Typing "  " when <p style="white-space:normal">A[]B</p>
  assert_in_array: value "<p style=\\"white-space:normal\\">A  B</p>" not in array ["<p style=\\"white-space:normal\\">A &nbsp;B</p>", "<p style=\\"white-space:normal\\">A&nbsp; B</p>"]
[FAIL] execCommand("insertText", false, "  ") when <p style="white-space:pre-line">A[]B</p>
  assert_in_array: value "<p style=\\"white-space:pre-line\\">A  B</p>" not in array ["<p style=\\"white-space:pre-line\\">A &nbsp;B</p>", "<p style=\\"white-space:pre-line\\">A&nbsp; B</p>"]
[FAIL] Typing "  " when <p style="white-space:pre-line">A[]B</p>
  assert_in_array: value "<p style=\\"white-space:pre-line\\">A  B</p>" not in array ["<p style=\\"white-space:pre-line\\">A &nbsp;B</p>", "<p style=\\"white-space:pre-line\\">A&nbsp; B</p>"]
[FAIL] execCommand("insertText", false, "a") when <p><b>[AB]</b></p> and execCommand("insertParagraph")
  assert_in_array: value "<p><br></p><p style=\\"font-weight: 700;\\">a</p>" not in array ["<p><b><br></b></p><p><b>a</b></p>", "<p><b><br></b></p><p><b>a<br></b></p>"]
[FAIL] Typing "a" when <p><b>[AB]</b></p> and execCommand("insertParagraph")
  assert_in_array: value "<p><br></p><p style=\\"font-weight: 700;\\">a</p>" not in array ["<p><b><br></b></p><p><b>a</b></p>", "<p><b><br></b></p><p><b>a<br></b></p>"]
[FAIL] execCommand("insertText", false, "a") when <p><b>[]AB</b></p> and execCommand("insertLineBreak")
  assert_equals: expected "<p><b><br>aAB</b></p>" but got "<p><b>\\naAB</b></p>"
[FAIL] Typing "a" when <p><b>[]AB</b></p> and execCommand("insertLineBreak")
  assert_equals: expected "<p><b><br>aAB</b></p>" but got "<p><b>\\naAB</b></p>"
[FAIL] execCommand("insertText", false, "a") when <p><b>A[]B</b></p> and execCommand("insertLineBreak")
  assert_equals: expected "<p><b>A<br>aB</b></p>" but got "<p><b>A\\naB</b></p>"
[FAIL] Typing "a" when <p><b>A[]B</b></p> and execCommand("insertLineBreak")
  assert_equals: expected "<p><b>A<br>aB</b></p>" but got "<p><b>A\\naB</b></p>"
[FAIL] execCommand("insertText", false, "a") when <p><b>AB[]</b></p> and execCommand("insertLineBreak")
  assert_in_array: value "<p><b>AB\\na</b></p>" not in array ["<p><b>AB<br>a</b></p>", "<p><b>AB<br>a<br></b></p>"]
[FAIL] Typing "a" when <p><b>AB[]</b></p> and execCommand("insertLineBreak")
  assert_in_array: value "<p><b>AB\\na</b></p>" not in array ["<p><b>AB<br>a</b></p>", "<p><b>AB<br>a<br></b></p>"]
[FAIL] execCommand("insertText", false, "a") when <p><b>[AB]</b></p> and execCommand("insertLineBreak")
  assert_in_array: value "<p>\\n<b>a</b></p>" not in array ["<p><b><br>a</b></p>", "<p><b><br>a<br></b></p>"]
[FAIL] Typing "a" when <p><b>[AB]</b></p> and execCommand("insertLineBreak")
  assert_in_array: value "<p>\\n<b>a</b></p>" not in array ["<p><b><br>a</b></p>", "<p><b><br>a<br></b></p>"]
[FAIL] execCommand("insertText", false, "a") when <p><b>[AB]</b></p> and execCommand("delete")
  assert_in_array: value "<p style=\\"font-weight: 700;\\">a</p>" not in array ["<p><b>a</b></p>", "<p><b>a<br></b></p>"]
[FAIL] Typing "a" when <p><b>[AB]</b></p> and execCommand("delete")
  assert_in_array: value "<p style=\\"font-weight: 700;\\">a</p>" not in array ["<p><b>a</b></p>", "<p><b>a<br></b></p>"]
[FAIL] execCommand("insertText", false, "a") when <p><b>A[]</b></p><p>B</p> and execCommand("delete") and move caret to the following paragraph temporarily
  assert_in_array: value "<p>a</p><p>B</p>" not in array ["<p><b>a</b></p><p>B</p>", "<p><b>a<br></b></p><p>B</p>"]
[FAIL] Typing "a" when <p><b>A[]</b></p><p>B</p> and execCommand("delete") and move caret to the following paragraph temporarily
  assert_in_array: value "<p>a</p><p>B</p>" not in array ["<p><b>a</b></p><p>B</p>", "<p><b>a<br></b></p><p>B</p>"]
[FAIL] execCommand("insertText", false, "a") when <p><b>[]A</b></p><p>B</p> and execCommand("forwardDelete") and move caret to the following paragraph temporarily
  assert_in_array: value "<p>a</p><p>B</p>" not in array ["<p><b>a</b></p><p>B</p>", "<p><b>a<br></b></p><p>B</p>"]
[FAIL] Typing "a" when <p><b>[]A</b></p><p>B</p> and execCommand("forwardDelete") and move caret to the following paragraph temporarily
  assert_in_array: value "<p>a</p><p>B</p>" not in array ["<p><b>a</b></p><p>B</p>", "<p><b>a<br></b></p><p>B</p>"]
[FAIL] execCommand("insertText", false, "a") when <p><b>[A]</b></p><p>B</p> and execCommand("delete") and move caret to the following paragraph temporarily
  assert_in_array: value "<p>a</p><p>B</p>" not in array ["<p><b>a</b></p><p>B</p>", "<p><b>a<br></b></p><p>B</p>"]
[FAIL] Typing "a" when <p><b>[A]</b></p><p>B</p> and execCommand("delete") and move caret to the following paragraph temporarily
  assert_in_array: value "<p>a</p><p>B</p>" not in array ["<p><b>a</b></p><p>B</p>", "<p><b>a<br></b></p><p>B</p>"]
Harness: the test ran to completion.
